requirements.txt
numpy==1.16.4
Keras==2.2.4
matplotlib==3.1.0
tensorflow==1.14.0
Pillow==6.1.0
